# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Carga Pesada",
        "description": "API RESTFul Carga Pesada",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/drivers": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Listar todos os motoristas",
                "operationId": "4cea0d334de40c00c0b375e65126e7e5",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtrar por nome",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cpf",
                        "in": "query",
                        "description": "Filtrar por cpf",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de motoristas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Driver"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Criar um novo motorista",
                "operationId": "d58b909c3a3eb72be38ffe5f7a7315e4",
                "requestBody": {
                    "description": "Dados do novo motorista",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateDriver"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorista criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Motorista criado com sucesso."
                                        },
                                        "driver": {
                                            "$ref": "#/components/schemas/Driver"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/drivers/{id}": {
            "get": {
                "tags": [
                    "Driver"
                ],
                "summary": "Obter detalhes de um motorista específico",
                "operationId": "40eeb1f7fdd478c840f3c42d22b2baed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do motorista",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do motorista",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Driver"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Motorista não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Motorista não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Driver": {
                "required": [
                    "id",
                    "nome",
                    "email",
                    "cpf",
                    "rg",
                    "data_nascimento",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do motorista (UUID)",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "description": "Nome",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF",
                        "type": "string"
                    },
                    "rg": {
                        "description": "RG",
                        "type": "string"
                    },
                    "data_nascimento": {
                        "description": "Data de Nascimento",
                        "type": "string",
                        "format": "date"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "StoreUpdateDriver": {
                "title": "StoreUpdateDriver",
                "required": [
                    "nome",
                    "email",
                    "cpf",
                    "rg",
                    "data_nascimento",
                    "password"
                ],
                "properties": {
                    "nome": {
                        "description": "Nome",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF",
                        "type": "string"
                    },
                    "rg": {
                        "description": "RG",
                        "type": "string"
                    },
                    "data_nascimento": {
                        "description": "Data de Nascimento",
                        "type": "string",
                        "format": "date"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}